T2I_Models/SAFREE/requirements.txt
albumentations==1.4.14
clip==1.0
diffusers==0.29.0
einops==0.3.0
ftfy==6.2.3
horovod==0.28.1
huggingface_hub==0.23.4
numpy==1.24.4
onnxruntime==1.18.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.0.0
Pillow==9.5.0
Pillow==10.4.0
progressbar33==2.4
pydload==1.0.9
python_sld==1.0.10
regex==2024.5.15
skimage==0.0
sld==0.0.1
stable_diffusion==0.0.0
timm==1.0.9
torch==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.65.0
transformers==4.44.1
